[Terrain] Terrarium format fixes (#8534)
* Change GSI presets to preserve the input color values.
Since these images are being used as gradient values, not as displayed colors, the authored values are the ones we'll want to continue using at runtime. In particular, for Terrarium-style images, changes to the color values can drastically change the encoded height values.

Signed-off-by: Mike Balfour <82224783+mbalfour-amzn@users.noreply.github.com>

* Fix the scaling formulas and add support for scaling of Terrarium images.

Signed-off-by: Mike Balfour <82224783+mbalfour-amzn@users.noreply.github.com>

* Fix the unit tests to properly test scaling and Terrarium data.
Also simplified the channel comparison logic so that tests can explicitly specify specific input pixel values and output gradient values to compare, instead of implicitly calculating them deeper in the helper methods.

Signed-off-by: Mike Balfour <82224783+mbalfour-amzn@users.noreply.github.com>

* Adjusted cube start positions for automated test to move them above terrain.
The fixes to gradient scaling caused the cubes to start below the terrain. They've been moved above the terrain once again, and the timeout for glass becoming stationary is larger to account for the further distance that it now travels.

Signed-off-by: Mike Balfour <82224783+mbalfour-amzn@users.noreply.github.com>