[Terrain] Adding support for bicubic filtering to image gradients. (#12800)
This PR adds bicubic filtering support for image gradients. While this could be useful in any application that needs to sample an image gradient at a higher resolution than the source image, this change was made mostly with the intention of improving terrain quality and allowing low resolution heightmaps to be used to make highly detailed terrain.