Make editing during an image reload more stable. (#13025)
If an image reload happens (from the user saving the image in a paint program for example) while still in image edit mode, instead of asserting and/or crashing, the Editor now replaces the paint program with the new data and lets the user keep painting.
The one quirky part is that any existing undo/redo buffers will no longer directly match the modified image, so they can introduce unwanted artifacts if the user tries to undo/redo painted data into an image that has been reloaded this way. Not perfect, but it's a better experience than crashing. :)

Signed-off-by: Mike Balfour <82224783+mbalfour-amzn@users.noreply.github.com>

Signed-off-by: Mike Balfour <82224783+mbalfour-amzn@users.noreply.github.com>